refactor(API): wip: register upload routes conditionally based on feature flags
