+++ api/src/routes/index.js
@@ -4,7 +4,9 @@ const { authenticate } = require('../middleware/auth');
 
 const router = express.Router();
 
-router.get('/health', (req, res) => { res.send('OK'); });
+router.get('/health', (req, res) => {
+  res.send('OK');
+});
 router.use('/auth', require('./auth/index'));
 router.use('/reports', require('./reports'));
 router.use('/about', require('./about'));
